Move announcemnt utils under browser-scripts and add a test for it
